integrated SocialMediaShareDialog in SocialMediaShareButton
