+++ frontend/src/components/project/ProjectPageRoot.js
@@ -21,7 +21,6 @@ import ProjectTeamContent from "./ProjectTeamContent";
 import { useLongPress } from "use-long-press";
 import { NOTIFICATION_TYPES } from "../communication/notifications/Notification";
 import SocialMediaShareButton from "./Buttons/SocialMediaShareButton";
-import SocialMediaShareDialog from "../dialogs/SocialMediaShareDialog";
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -422,14 +421,10 @@ export default function ProjectPageRoot({
           </Tabs>
         </div>
         {!screenSize.belowSmall && (
-          <>
             <SocialMediaShareButton
               containerClassName={classes.shareButtonContainer}
               toggleShowSocials={toggleShowSocials}
-            />
-            <SocialMediaShareDialog
-              open={showSocials}
-              onClose={toggleShowSocials}
+              showSocials={showSocials}
               texts={texts}
               project={project}
               locale={locale}
@@ -437,7 +432,6 @@ export default function ProjectPageRoot({
               createShareRecord={createShareRecord}
               screenSize={screenSize}
             />
-          </>
         )}
       </Container>
